updated delayed job instrumentation
