+++ test/multiverse/suites/delayed_job/Envfile
@@ -1,3 +1,5 @@
+instrumentation_methods :chain, :prepend
+
 boilerplate_gems = <<-SQLITE
   gem 'rack'
   if RUBY_PLATFORM == 'java'
