+++ test/multiverse/suites/delayed_job/config/newrelic.yml
@@ -5,6 +5,8 @@ development:
   apdex_t: 0.5
   monitor_mode: true
   license_key: bootstrap_newrelic_admin_license_key_000
+  instrumentation:
+      delayed_job: <%= $instrumentation_method %>
   app_name: test
   host: 127.0.0.1
   api_host: 127.0.0.1
